Move infer::region_constraints::MemberConstraint to infer::types module.
